Add Button to open the Map to Stop Main Page
Refactored a similar button from Routes Main Page to a generic
component.
