+++ ui/src/components/stop-registry/main/StopRegistryMainPage.tsx
@@ -1,7 +1,8 @@
 import { useTranslation } from 'react-i18next';
 import { Container, Row } from '../../../layoutComponents';
 import { SimpleButton } from '../../../uiComponents';
-import { StopSearchBar } from '../search/StopSearchBar';
+import { OpenDefaultMapButton } from '../../common/OpenDefaultMapButton';
+import { StopSearchBar } from '../search';
 
 const testIds = {
   createStopButton: 'StopRegistryMainPage::createStopButton',
@@ -14,17 +15,17 @@ export const StopRegistryMainPage = (): JSX.Element => {
     <Container>
       <Row className="justify-between">
         <h1>{t('stops.stops')}</h1>
-        <div className="space-x-4">
-          <SimpleButton
-            // TODO: actual implementation for this button.
-            id="create-stop-button"
-            testId={testIds.createStopButton}
-            href="/"
-            disabled
-          >
-            {t('stops.createStop')}
-          </SimpleButton>
-        </div>
+        <OpenDefaultMapButton containerClassName="ml-auto" />
+        <SimpleButton
+          // TODO: actual implementation for this button.
+          id="create-stop-button"
+          testId={testIds.createStopButton}
+          href="/"
+          disabled
+          containerClassName="ml-3"
+        >
+          {t('stops.createStop')}
+        </SimpleButton>
       </Row>
       <StopSearchBar />
     </Container>
